refactor(profiling): extract interrupt manager
